Fixed error in excel reports, added migration to add 2006 exchange rates and added new_relic plugin
git-svn-id: file:///svn/trunk/odadata@8737 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ lib/report/excel.rb
@@ -54,7 +54,7 @@ module Report
       row.map do |cell| 
         # If this is a currency object convert to string without unit identifier
         if cell.is_a?(MultiCurrency::ConvertibleCurrency)
-          str_repr = cell.to_f
+          str_repr = cell.to_i
         else
           str_repr = cell.to_s
           #str_repr = cell.is_a?(MultiCurrency::ConvertibleCurrency) ? cell.to_s(false) : cell.to_s
